requirements-test.txt
coverage>=7.6.0
flake8>=7.0.0
Flask>=2.0.2
mysql-connector-python>=8.2.0
pydantic>=2.10.0
pytest>=8.3.0
